[fe#37] rename: 폴더구조 세분화
