+++ fe/src/App.js
@@ -1,7 +1,7 @@
-import Header from './views/Header.js';
 import { Reset } from 'styled-reset';
-import { BestProduct } from './views/BestProduct.js';
-import { Category } from './views/Category';
+import Header from './views/header/Header.js';
+import { BestProduct } from './views/main/bestProduct/BestProduct.js';
+import { Category } from './views/main/category/Category.js';
 
 function App() {
   return (
