+++ fe/src/views/main/bestProduct/tabList/tabList.js
@@ -1,5 +1,5 @@
-import { StyledDiv } from './BestProduct.styled';
-import { ProductCard } from '../../components/ProductCard';
+import { StyledDiv } from '../BestProduct.styled';
+import { ProductCard } from '../../../../components/ProductCard';
 
 export function TabList({ tabList }) {
   if (!tabList) {
