+++ fe/src/views/main/bestProduct/tabMenu/tabMenu.js
@@ -1,4 +1,4 @@
-import { TabMenuLi } from './BestProduct.styled';
+import { TabMenuLi } from '../BestProduct.styled';
 
 export function TabMenu(props) {
   if (!props.menus) {
